requirements.txt
cycler==0.11.0
fonttools==4.33.3
imageio==2.19.2
kiwisolver==1.4.2
matplotlib==3.5.2
networkx==2.8
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
tabulate==0.8.9
tifffile==2022.5.4
websocket-client==1.2.3
